Return original code
