--- xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisTickCalculatorType.java
@@ -1,9 +1,0 @@
-package org.knowm.xchart.internal.chartpart;
-
-public enum AxisTickCalculatorType {
-  AxisTickCalculator_Callback,
-  AxisTickCalculator_Category,
-  AxisTickCalculator_Date,
-  AxisTickCalculator_Logarithmic,
-  AxisTickCalculator_Number
-}
