+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisTickCalculator_Date.java
@@ -2,7 +2,6 @@ package org.knowm.xchart.internal.chartpart;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.knowm.xchart.internal.Utils;
@@ -27,7 +26,7 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
   private static final long YEAR_SCALE = TimeUnit.DAYS.toMillis(1L) * 365;
 
   private static final List<TimeSpan> timeSpans = new ArrayList<>();
-  private static List<Double> nullList = Collections.emptyList();
+
   static {
     timeSpans.add(new TimeSpan(MILLIS_SCALE, 1, "ss.SSS"));
     timeSpans.add(new TimeSpan(MILLIS_SCALE, 2, "ss.SSS"));
@@ -99,7 +98,7 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
       double maxValue,
       AxesChartStyler styler) {
 
-    super(axisDirection, workingSpace, minValue, maxValue, nullList, styler);
+    super(axisDirection, workingSpace, minValue, maxValue, styler);
 
     calculate();
   }
@@ -140,6 +139,7 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
 
     // the span of the data
     long span = (long) Math.abs(maxValue - minValue); // in data space
+    // System.out.println("span: " + span);
 
     // Generate the labels first, see if they "look" OK and reiterate with an increased
     // tickSpacingHint
@@ -150,7 +150,9 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
             - 5;
     int gridStepInChartSpace;
 
+    // System.out.println("calculating ticks...");
     long gridStepHint = (long) (span / tickSpace * tickSpacingHint); // in time units (ms)
+    // System.out.println("gridStepHint: " + gridStepHint);
 
     //////////////////////////////////////////////
 
@@ -166,9 +168,12 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
         break;
       }
     }
+    //    TimeSpan timeSpan1 = timeSpans.get(index);
+    //    System.out.println("timeSpan1 = " + timeSpan1);
 
     // use the pattern from the first timeSpan
     String datePattern = timeSpans.get(index).getDatePattern();
+    // System.out.println("index: " + index);
 
     // iterate BACKWARDS from previous point until the appropriate timespan is found for the
     // gridStepHint
@@ -201,7 +206,14 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
         skip = true;
         continue;
       }
+      //      TimeSpan timeSpan2 = timeSpans.get(index);
+      //      System.out.println("timeSpan2 = " + timeSpan2);
+
+      // System.out.println("gridStepInChartSpace: " + gridStepInChartSpace);
+
       double firstPosition = getFirstPosition(gridStep);
+      //      System.out.println("firstPosition = " + firstPosition);
+      //      System.out.println("   " + new Date((long) firstPosition).toGMTString());
 
       // Define Date Pattern
       // override pattern if one was explicitly given
@@ -212,6 +224,7 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
 
       SimpleDateFormat simpleDateformat = new SimpleDateFormat(datePattern, styler.getLocale());
       simpleDateformat.setTimeZone(styler.getTimezone());
+      //      simpleDateformat.setTimeZone(TimeZone.getTimeZone("UTC"));
       axisFormat = simpleDateformat;
 
       // generate all tickLabels and tickLocations from the first to last position
@@ -220,13 +233,19 @@ class AxisTickCalculator_Date extends AxisTickCalculator_ {
           value = value + gridStep) {
 
         tickLabels.add(axisFormat.format(value));
+        //        System.out.println("ticklabel date = " + new Date((long) value).toGMTString());
+        // here we convert tickPosition finally to plot space, i.e. pixels
         double tickLabelPosition =
             margin + ((value - minValue) / (maxValue - minValue) * tickSpace);
+        // System.out.println("tickLabelPosition: " + tickLabelPosition);
         tickLocations.add(tickLabelPosition);
+        // }
       }
+      //      System.out.println("************");
     } while (skip
         || !areAllTickLabelsUnique(tickLabels)
         || !willLabelsFitInTickSpaceHint(tickLabels, gridStepInChartSpace));
+    //    System.out.println("are ticklabels unique? " + areAllTickLabelsUnique(tickLabels));
   }
 
   static class TimeSpan {
