+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisTickCalculator_Override.java
@@ -1,8 +1,6 @@
 package org.knowm.xchart.internal.chartpart;
 
 import java.text.SimpleDateFormat;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import org.knowm.xchart.internal.Utils;
@@ -15,8 +13,7 @@ import org.knowm.xchart.style.AxesChartStyler;
  * rendering the axis ticks for given values & labels
  */
 class AxisTickCalculator_Override extends AxisTickCalculator_ {
-  private static List<Double> nullList = Collections.emptyList();
-  private Map<Double, Object> labelOverrideMap;
+
   /**
    * Constructor for Numerical axis
    *
@@ -26,9 +23,6 @@ class AxisTickCalculator_Override extends AxisTickCalculator_ {
    * @param maxValue
    * @param styler
    * @param labelOverrideMap
-   * @param markMap
-   * @param axisType
-   * @param categoryCount
    */
   public AxisTickCalculator_Override(
       Direction axisDirection,
@@ -36,16 +30,33 @@ class AxisTickCalculator_Override extends AxisTickCalculator_ {
       double minValue,
       double maxValue,
       AxesChartStyler styler,
-      Map<Double, Object> labelOverrideMap,
+      Map<Double, Object> labelOverrideMap) {
+
+    super(axisDirection, workingSpace, minValue, maxValue, styler);
+    axisFormat = new Formatter_Number(styler, axisDirection, minValue, maxValue);
+    calculate(labelOverrideMap);
+  }
+
+  /**
+   * Constructor for category axis
+   *
+   * @param axisDirection
+   * @param workingSpace
+   * @param styler
+   * @param markMap
+   * @param axisType
+   * @param categoryCount
+   */
+  public AxisTickCalculator_Override(
+      Direction axisDirection,
+      double workingSpace,
+      AxesChartStyler styler,
       Map<Double, Object> markMap,
       Series.DataType axisType,
       int categoryCount) {
 
-    super(axisDirection, workingSpace, minValue, maxValue, nullList, styler);
-    this.labelOverrideMap = labelOverrideMap;
-    axisFormat = new Formatter_Number(styler, axisDirection, minValue, maxValue);
-    calculate();
-    
+    super(axisDirection, workingSpace, Double.NaN, Double.NaN, styler);
+
     // set up String formatters that may be encountered
     if (axisType == Series.DataType.String) {
       axisFormat = new Formatter_String();
@@ -60,12 +71,11 @@ class AxisTickCalculator_Override extends AxisTickCalculator_ {
       simpleDateformat.setTimeZone(styler.getTimezone());
       axisFormat = simpleDateformat;
     }
-    
+
     calculateForCategory(markMap, categoryCount);
   }
-  
-  @Override
-  protected void calculate() {
+
+  private void calculate(Map<Double, Object> labelOverrideMap) {
 
     // a check if all axis data are the exact same values
     if (minValue == maxValue) {
